chore: Add types for ag-grid-react

+++ src/book-disposal/entity/book-disposal.entity.ts
@@ -3,7 +3,7 @@ import { Column, Entity } from 'typeorm';
 
 @Entity()
 export class BookDisposalModel extends BaseModel {
-  @Column({ type: 'int', comment: '번호', nullable: true })
+  @Column({ type: 'float', comment: '번호', nullable: true })
   no: number;
 
   @Column({ type: 'date', comment: '날짜', nullable: true })
@@ -33,22 +33,22 @@ export class BookDisposalModel extends BaseModel {
   @Column({ type: 'text', comment: '마크장비', nullable: true })
   mark_equip: string;
 
-  @Column({ type: 'int', comment: '최종납품권수', nullable: true })
+  @Column({ type: 'float', comment: '최종납품권수', nullable: true })
   final_delivery_bks: number;
 
-  @Column({ type: 'int', comment: '최종매출액', nullable: true })
+  @Column({ type: 'float', comment: '최종매출액', nullable: true })
   final_sales: number;
 
-  @Column({ type: 'int', comment: '지출비용', nullable: true })
+  @Column({ type: 'float', comment: '지출비용', nullable: true })
   expend_cost: number;
 
   @Column({ type: 'date', comment: '선입금일자', nullable: true })
   pre_payment_date: Date;
 
-  @Column({ type: 'int', comment: '선입금', nullable: true })
+  @Column({ type: 'float', comment: '선입금', nullable: true })
   pre_payment: number;
 
-  @Column({ type: 'int', comment: '예정잔금', nullable: true })
+  @Column({ type: 'float', comment: '예정잔금', nullable: true })
   expected_balance: number;
 
   @Column({ type: 'text', comment: '업체연락처', nullable: true })
